fix: sort all data not page
